requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
python-dotenv==1.0.0
pydantic==2.5.3
email-validator==2.1.0
pydantic-settings==2.1.0
python-multipart==0.0.6
alembic==1.13.1
slowapi==0.1.9
redis==5.0.1
bleach==6.1.0
google-api-python-client==2.118.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
apscheduler==3.10.4
cryptography==42.0.2
httpx==0.27.0
